+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX0BE9WJWVX1SCGJF6DYXNK3"
+        "01KX0VJTKYCM543VR6V5D4RN5D"
     )
     r"""The unique identifier of the data source"""