+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-08T07:52:34.825Z"
+    created: Optional[str] = "2026-07-08T12:34:34.233Z"
 
-    updated: Optional[str] = "2026-07-08T07:52:34.825Z"
+    updated: Optional[str] = "2026-07-08T12:34:34.233Z"
 
     updated_by_id: OptionalNullable[str] = UNSET