+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX0BE9WE2SZCY1B640FDAPFT"
+        "01KX0VJTKS34992S25EEJ0V57D"
     )
     r"""The unique identifier of the data source"""