+++ packages/orq-rc/RELEASES.md
@@ -9349,3 +9349,13 @@ Based on:
 - [python v4.12.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.7 - packages/orq-rc
+
+## 2026-07-08 12:37:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.8 - packages/orq-rc